requirements.txt
azure-functions
requests
azure-eventhub